Change PixelDraw(..., INT8* pScreen) to PixelDraw(..., UINT16* pScreen).
git-svn-id: svn://tron.homeunix.org/ja2/trunk@3492 5e31c081-6ce3-0310-bb30-f584a8092234

+++ sgp/Line.h
@@ -24,7 +24,7 @@ void SetClippingRegionAndImageWidth(int iImageWidth, int iClipStartX, int iClipS
 //	Don't send fClip==TRUE to LineDraw if you don't have to. So if you know
 //  that your line will be within the region you want it to be in, set
 //	fClip == FALSE.
-void PixelDraw( BOOLEAN fClip, INT32 xp, INT32 yp, INT16 sColor, INT8 *pScreen );
+void PixelDraw(BOOLEAN fClip, INT32 xp, INT32 yp, INT16 sColor, UINT16* pScreen);
 void LineDraw(BOOLEAN fClip, int XStart, int YStart, int XEnd, int YEnd, short Color, char* ScreenPtr);
 void RectangleDraw(BOOLEAN fClip, int XStart, int YStart, int XEnd, int YEnd, short Color, char* ScreenPtr);
